test: move test/setup & teardown into test/suite.js
- so that running 'node --test' doesn't to teardown during test suite runs
