+++ lib/user.js
@@ -6,6 +6,8 @@ const Util = require('./util')
 Util.setEnv()
 const Config = require('./config')
 
+const userDbMap = { id: 'nt_user_id', gid: 'nt_group_id' }
+
 class User {
   constructor(args) {
     this.debug = args?.debug ?? false
@@ -92,7 +94,7 @@ class User {
       , username
       , email
      FROM nt_user WHERE`,
-      Util.mapToDbColumn(args, { id: 'nt_user_id', gid: 'nt_group_id' }),
+      Util.mapToDbColumn(args, userDbMap),
     )
   }
 
@@ -108,22 +110,22 @@ class User {
       , email
       , deleted
      FROM nt_user WHERE`,
-      Util.mapToDbColumn(args, { id: 'nt_user_id', gid: 'nt_group_id' }),
+      Util.mapToDbColumn(args, userDbMap),
     )
   }
 
   async delete(args, val) {
-    const u = await this.getAdmin({ nt_user_id: args.nt_user_id })
-    if (u.length === 1) {
-      await Mysql.execute(`UPDATE nt_user SET deleted=? WHERE nt_user_id=?`, [
-        val ?? 1,
-        u[0].id,
-      ])
-    }
+    const u = await this.getAdmin(args)
+    if (u.length !== 1) return false
+    await Mysql.execute(`UPDATE nt_user SET deleted=? WHERE nt_user_id=?`, [
+      val ?? 1,
+      u[0].id,
+    ])
+    return true
   }
 
   async destroy(args) {
-    const u = await this.getAdmin({ nt_user_id: args.nt_user_id })
+    const u = await this.getAdmin(args)
     if (u.length === 1) {
       await Mysql.execute(`DELETE FROM nt_user WHERE nt_user_id=?`, [u[0].id])
     }
