+++ lib/user.test.js
@@ -38,7 +38,7 @@ describe('user', function () {
     })
 
     it('deletes a user', async () => {
-      await user.delete({ nt_user_id: 4096 })
+      assert.ok(await user.delete({ nt_user_id: 4096 }))
       let u = await user.getAdmin({ nt_user_id: 4096 })
       assert.equal(u[0].deleted, 1)
       await user.delete({ nt_user_id: 4096 }, 0) // restore
